Fixed cases of undefined behavior if the std::optional like memory layout was used.
The issue was that the object of type SeparateFlagWrapper was never created. We only popped
its two members into existence (separately), more or less, but since SeparateFlagWrapper itself
was never created, this was UB.

Moreover, SeparateFlagWrapper::isEmptyFlag was never actually created, too. It is just a bool,
but a bool should have been created nevertheless. Hence, OrdinaryBoolFlagManipulator
operated on a bool that was never really created.

Now SeparateFlagWrapper gets created together with the optional, and the isEmptyFlag is an
ordinary member that gets created at the same time.

Note that this commit is all about creating and destroying C++ objects. They get created with
indeterminate values. It is not about assigning a specific value to the objects.