Added TINY_OPTIONAL_VERSION and put everything into an inline namespace.
Mixing different versions and configurations of the library is bad. They
are not ABI compatiable in general. To prevent unintended mixing, this
commit puts everything into an inline namespace. Except
tiny::optional_flag_manipulator, because that would require that the
optional_flag_manipulator_fwd.h file knows about the version and
configuration. This is too cumbersome to do.

To implement the inline namespace properly, also added
TINY_OPTIONAL_VERSION which allows to identify the version of the
library.