Amended Natvis to account for the new inline namespace.
Unfortunately, Natvis needs the inline namespace to resolve the types.
This also means that both the version number and the configuration gets
hardcoded in Natvis. I think there is no way around this. (We have the
same problem with e.g. the abseil library.)
So at least explain it in the Natvis and the readme.