Made versions with same major and minor but different patch versions compatible to each other + incremented version to 1.4.0.
I.e. we guarantee compatibility between versions of the library with different patch versions if the major and minor versions are the same. This allows to make releases with just e.g. documentation or test updates, where the patch version gets incremented and now no longer leads to a breaking change.