Refactoring: Replace the conditional inheritance hierarchy with `requires` overloads in C++20
In C++17, the only option to satisfy the conditional deletion of the copy/move constructor/assignment was via conditional inheritance. This leads to pretty complex code, and it is also quite slow to compile.
In C++20 we can simply attach some `requires` clauses and overload the constructors/assignment operators via them. This simplifies the code and it is faster to compile, too. It was like ~20%-40% faster than the C++17 version in my tests.

Note: Triviality is so far only implemented for the move constructor.