Simplified the conditions for making the copy/move constructors/assignment operators trivial.
I realized that we do not need to take into account the possibility that padding bytes might be exploited to store the IsEmpty flag. Either the they are not exploited and we are fine. Or they are exploited and then the constructors/assignments necessarily need to be non-trivial, in which case we are fine too.
In other words: Removed the FlagManipulator::tiny_optional_allow_trivial_move_copy and IsStandardTypeAllowingTrivialMoveCopy stuff.
-> We can follow the standard for std::optional to 100% regarding the conditions about triviality.

Also updated readme. Especially added a warning about padding bytes.